fixed type error in network interface manager

+++ rocon_gateway/src/rocon_gateway/network_interface_manager.py
@@ -105,9 +105,9 @@ class NetworkInterfaceManager(object):
             return gateway_statistics
 
         wifi = pythonwifi.Wireless(self.interface_name)
-        gateway_statistics.wireless_bitrate = wifi.wireless_info.getBitrate().value #Raw bitrate
+        gateway_statistics.wireless_bitrate = float(wifi.wireless_info.getBitrate().value) #Raw bitrate
         _, qual, _, _ = wifi.getStatistics()
-        gateway_statistics.wireless_link_quality = float(qual.quality)
+        gateway_statistics.wireless_link_quality = int(qual.quality)
         gateway_statistics.wireless_signal_level = float(qual.signallevel) - 256.0 # HACK. something seems to be wrong with the value returned from pythonwifi
         gateway_statistics.wireless_noise_level = float(qual.noiselevel)
